requirements.txt
streamlit
pytesseract
opencv-python
pillow
pandas
easyocr
opencv-python